Глайдер, управление и слежение за ним.

+++ Engine/Draw/Camera.cpp
@@ -172,7 +172,7 @@ void Camera::rotate(const vec2& angles)
 {
 	vec3 vector = _vec;
 	double angleY = asinf(vector.z);
-	double angleX = acosf(vector.y / cosf(angleY));
+	double angleX = acos(vector.y / cos(angleY));
 
 	if (vector.x < 0.0f)
 	{
@@ -185,9 +185,9 @@ void Camera::rotate(const vec2& angles)
 	if (angleY >(glm::pi<float>() / 2.0 - 0.25)) angleY = (glm::pi<float>() / 2.0 - 0.25);
 	if (angleY < -(glm::pi<float>() / 2.0 - 0.25)) angleY = -(glm::pi<float>() / 2.0 - 0.25);
 
-	vector.x = sinf(angleX) * cosf(angleY);
-	vector.y = cosf(angleX) * cosf(angleY);
-	vector.z = sinf(angleY);
+	vector.x = sinf(static_cast<float>(angleX)) * cosf(static_cast<float>(angleY));
+	vector.y = cosf(static_cast<float>(angleX)) * cosf(static_cast<float>(angleY));
+	vector.z = sinf(static_cast<float>(angleY));
 	normalize(vector);
 
 	setVector(vector);
